add delete method and getAll method for Department

+++ src/main/java/sa/aref/university/service/department/DepartmentServiceImpl.java
@@ -43,7 +43,12 @@ public class DepartmentServiceImpl implements DepartmentService {
     }
 
     @Override
-    public void deleteDepartmentById(Long id) {
-        departmentRepository.deleteById(id);
+    public boolean deleteDepartmentById(Long id) {
+        Department department = departmentRepository.findById(id).orElse(null);
+        if (department != null) {
+            departmentRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
